requirements.txt
asyncio
discord.py
praw
aiosqlite
dblpy
DiscordUtils
jishaku
aiomojang
youtube_dl